backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.1",
    "jinja2>=3.1.6",
    "passlib[argon2,bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
